+++ CHANGELOG.md
@@ -4,12 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
+- (none)
+
+## [1.1.0] - 2016-06-16
 ### Added
 - Added [`no-delete`] rule
 - Added [`no-nil`] rule
 - Added [`no-unused-expression`] rule
 
-## [1.0.0] - 2016-06-15
+## 1.0.0 - 2016-06-15
 ### Added
 - Added [`no-class`] rule
 - Added [`no-let`] rule
@@ -19,7 +22,8 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 - Added [`no-this`] rule
 - Added [`no-throw`] rule
 
-[Unreleased]: https://github.com/jfmengels/eslint-plugin-fp/compare/v1.0.0...master
+[Unreleased]: https://github.com/jfmengels/eslint-plugin-fp/compare/v1.1.0...master
+[1.1.0]: https://github.com/jfmengels/eslint-plugin-fp/compare/v1.0.0...v1.1.0
 
 [`no-class`]: ./docs/rules/no-class.md
 [`no-delete`]: ./docs/rules/no-delete.md